feat(api): rename user repositories folder
